backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
psycopg2-binary==2.9.9
firebase-admin==6.5.0
python-dotenv==1.0.1
gunicorn==22.0.0
Werkzeug==3.0.3
